collect AA/Nemo/Hecke stuff in Misc.jl
Move code to `Misc.jl` that belongs to AbstractAlgebra.jl,
Nemo.jl, or Hecke.jl.

Once we agree that this code should be moved to these packages,
the code should be added there,
and as soon as it becomes available in Oscar, it can be removed
from `Misc.jl`.
