+++ experimental/GModule/Brueckner.jl
@@ -3,17 +3,6 @@ using Oscar
 
 export coimage
 
-Base.pairs(M::MatElem) = Base.pairs(IndexCartesian(), M)
-Base.pairs(::IndexCartesian, M::MatElem) = Base.Iterators.Pairs(M, CartesianIndices(axes(M)))
-
-function Hecke.roots(a::FinFieldElem, i::Int)
-  kx, x = polynomial_ring(parent(a), cached = false)
-  return roots(x^i-a)
-end
-
-Oscar.matrix(phi::Generic.IdentityMap{<:AbstractAlgebra.FPModule}) = identity_matrix(base_ring(domain(phi)), dim(domain(phi)))
-
-
 #=TODO
  - construct characters along the way as well?
  - compare characters rather than the hom_base
@@ -276,68 +265,9 @@ function brueckner(mQ::Map{<:Oscar.GAPGroup, PcGroup}; primes::Vector=[])
   return allR
 end
 
-Oscar.gen(M::AbstractAlgebra.FPModule, i::Int) = M[i]
-
-Oscar.is_free(M::Generic.FreeModule) = true
-Oscar.is_free(M::Generic.DirectSumModule) = all(is_free, M.m)
-
-function Oscar.dual(h::Map{FinGenAbGroup, FinGenAbGroup})
-  A = domain(h)
-  B = codomain(h)
-  @assert is_free(A) && is_free(B)
-  return hom(B, A, transpose(h.map))
-end
-
-function Oscar.dual(h::Map{<:AbstractAlgebra.FPModule{ZZRingElem}, <:AbstractAlgebra.FPModule{ZZRingElem}})
-  A = domain(h)
-  B = codomain(h)
-  @assert is_free(A) && is_free(B)
-  return hom(B, A, transpose(matrix(h)))
-end
-
 function coimage(h::Map)
   return quo(domain(h), kernel(h)[1])
 end
-
-function Oscar.cokernel(h::Map)
-  return quo(codomain(h), image(h)[1])
-end
-
-function Base.iterate(M::AbstractAlgebra.FPModule{T}) where T <: FinFieldElem
-  k = base_ring(M)
-  if dim(M) == 0
-    return zero(M), iterate([1])
-  end
-  p = Base.Iterators.ProductIterator(Tuple([k for i=1:dim(M)]))
-  f = iterate(p)
-  return M(elem_type(k)[f[1][i] for i=1:dim(M)]), (f[2], p)
-end
-
-function Base.iterate(::AbstractAlgebra.FPModule{<:FinFieldElem}, ::Tuple{Int64, Int64})
-  return nothing
-end
-
-function Base.iterate(M::AbstractAlgebra.FPModule{T}, st::Tuple{<:Tuple, <:Base.Iterators.ProductIterator}) where T <: FinFieldElem
-  n = iterate(st[2], st[1])
-  if n === nothing
-    return n
-  end
-  return M(elem_type(base_ring(M))[n[1][i] for i=1:dim(M)]), (n[2], st[2])
-end
-
-function Base.length(M::AbstractAlgebra.FPModule{T}) where T <: FinFieldElem
-  return Int(order(base_ring(M))^dim(M))
-end
-
-function Base.eltype(M::AbstractAlgebra.FPModule{T}) where T <: FinFieldElem
-  return elem_type(M)
-end
-
-function Oscar.dim(M::AbstractAlgebra.Generic.DirectSumModule{<:FieldElem})
-  return sum(dim(x) for x = M.m)
-end
-
-Oscar.is_finite(M::AbstractAlgebra.FPModule{<:FinFieldElem}) = true
 
 """
   mp: G ->> Q
